Keep calendar adjustment transitions reviewable
Constraint: The feature arrived as a 1451-line domain module plus MCP glue in an existing monolith.
Rejected: Rewriting the state machine while extracting it | unnecessary behavioral risk before upstream review.
Confidence: high
Scope-risk: narrow
Directive: Preserve the adjustments facade and keep atomic transition code auditable as one service/store contract.
Tested: 162 targeted calendar, MCP, store, and Google adapter tests; ruff; compileall; sdist and wheel build.
Not-tested: Full suite after extraction is deferred to the final stacked-branch gate.
